FELIX-6121 : Support servlet 4.0
git-svn-id: https://svn.apache.org/repos/asf/felix/trunk@1858782 13f79535-47bb-0310-9956-ffa450edef68
