+++ http/proxy/pom.xml
@@ -42,6 +42,17 @@
             <plugin>
                 <groupId>org.apache.felix</groupId>
                 <artifactId>maven-bundle-plugin</artifactId>
+                <configuration>
+                    <instructions>
+                        <Import-Package>
+                            javax.servlet;version="[3.1,0)",
+                            javax.servlet.annotation;version="[3.1,0)",
+                            javax.servlet.descriptor;version="[3.1,0)",
+                            javax.servlet.http;version="[3.1,0)",
+                            *
+                        </Import-Package>
+                    </instructions>
+                </configuration>
             </plugin>
         </plugins>
     </build>
@@ -50,6 +61,7 @@
         <dependency>
             <groupId>javax.servlet</groupId>
             <artifactId>javax.servlet-api</artifactId>
+            <version>4.0.1</version>
         </dependency>
         <dependency>
             <groupId>org.osgi</groupId>
