+++ http/proxy/src/main/java/org/apache/felix/http/proxy/package-info.java
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-@Version("3.0.0")
+@Version("3.1.0")
 package org.apache.felix.http.proxy;
 
 import org.osgi.annotation.versioning.Version;
